src/requirements/django/requirements.txt
Django==5.0.6
psycopg[binary]==3.1.19
django-environ==0.11.2
django-cors-headers==4.3.1
pyjwt==2.8.0
oauthlib==3.2.2
